OT275-79: refactor mappers
